- fixed the single issue template
First, we removed the `link` option as the default. If you choose Custom
Text in the post then it will also try to send you to the link. When the
link isn't set (because you can't set it when you have picked custom
text) it will simply cycle you on the page with a blank URL value. Now
you only get a the link to an external resource as a redirect when you
choose external link.

Second, revamped the template in light of security. Revamping the
escaping and overall structure of the vars.