- updated so we do not try to print document info if no file attached
Previously we would try to print document info if there was content or
if there was a file. Since document info could exist in our content
blocks without a file, the file read would error as there was no array
defined.

Now we will show the content if it exists, but skip the file download if
it doesn't exist.